Tweaked some pronoun consistency
In section 2 and elsewhere, 'we' refers to developers who are using itty, or code written by them.

In sections 3 and 4, 'we' seemed to be referring to the itty router itself. This probabaly wouldn't affect anyone but me ... but it did affect me so here's a PR :)

I'm not sure my suggested changes are the best, and I'd be unoffended by modification and improvement.